Use the java duration objects instead of raw integers for timeouts and timestamps.
This might be ever so slightly slower, but I am using the duration
objects instead of raw-dogging integers.  This helps me avoid making
mistakes with unit conversions (e.g. mixing up nanoseconds and
milliseconds and seconds). It also makes it easier to work with
nanoseconds, which might be ever so slightly more precise. Not that that
matters much.