Fix a bug where restarting pauses updates forever.
There was an issue that if the swaybar is killed in the middle of
processing, it would never update because it would never get into the
update block.

Fortunately, I actually implemented a fix for this already meaning that
all I had to do was allow non-async stuff to reuse that logic.